pyproject.toml
[tool.poetry]
name = "internetshop"
version = "0.1.0"
description = "rest api based on django rest framework"
authors = ["TsMarch"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.7"
asgiref = "3.8.1"
black = "24.4.2"
click = "8.1.7"
colorama = "0.4.6"
django = "5.0.7"
django-types = "0.19.1"
djangorestframework = "3.15.2"
isort = "5.13.2"
mypy-extensions = "1.0.0"
packaging = "24.1"
pathspec = "0.12.1"
platformdirs = "4.2.2"
sqlparse = "0.5.1"
types-psycopg2 = "2.9.21.20240417"
tzdata = "2025.2"
django-filter = "24.3"
build = "1.2.2.post1"
cachecontrol = "0.14.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
cleo = "2.1.0"
crashtest = "0.4.1"
distlib = "0.3.9"
djangorestframework-simplejwt = "5.3.1"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.16.1"
idna = "3.10"
importlib-metadata = "8.5.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
more-itertools = "10.5.0"
msgpack = "1.1.0"
pexpect = "4.9.0"
pkginfo = "1.11.2"
poetry = "1.8.4"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pyjwt = "2.9.0"
pyproject-hooks = "1.2.0"
pywin32-ctypes = "0.2.3"
rapidfuzz = "3.10.0"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
urllib3 = "2.2.3"
virtualenv = "20.27.0"
zipp = "3.20.2"
cfgv = "3.4.0"
identify = "2.6.1"
nodeenv = "1.9.1"
pre-commit = "4.0.1"
pyyaml = "6.0.2"
flake8 = "7.1.1"
mccabe = "0.7.0"
pycodestyle = "2.12.1"
pyflakes = "3.2.0"
django-eav2 = "1.7.1"
psycopg2-binary = "2.9.10"
python-dotenv = "1.0.1"
openpyxl = "^3.1.5"
pandas = "2.2.3"
attrs = "25.1.0"
cffi = "1.17.1"
drf-spectacular = "0.28.0"
et-xmlfile = "2.0.0"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
numpy = "2.2.2"
python-dateutil = "2.9.0.post0"
pytz = "2024.2"
referencing = "0.36.2"
rpds-py = "0.22.3"
six = "1.17.0"
typing-extensions = "4.12.2"
uritemplate = "4.1.1"
celery = "5.5.1"
django-mptt = "0.16.0"
tqdm = "4.67.1"
django-extensions = "^4.1"
redis = "^5.2.1"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
skip = ["internet_shop/config/urls.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
